requirements.txt
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
decorator==4.4.2
filelock==3.14.0
fsspec==2024.6.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai-whisper==20231117
pillow==10.3.0
pocketsphinx==5.0.3
proglog==0.1.10
pycparser==2.22
pydub==0.25.1
regex==2024.5.15
requests==2.32.3
sounddevice==0.4.7
SpeechRecognition==3.10.4
sympy==1.12.1
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
